power: rk817-battery: support rk809
A “battery fuel gauge” is integrated in the RK809. Using
the proprietary algorithms and the sensed battery current
and voltage, the gauge can accurately calculate the battery
capacity based on the charging/discharging characteristics
of the battery preloaded in the system. The gauge then sends
the battery capacity information to the processor through
the I2C interface.

Change-Id: I0be1f9ef48285c17d599753d65161e79b0a4c95a
Signed-off-by: Shengfei Xu <xsf@rock-chips.com>

rockchip ef9512730326fd6c10d405c59a8d38d3854e5955